refactor(ime): extract pure overlay pixel-math into ime_overlay/layout.rs
